Allow to share MX queries between sessions on the same route.
When a session wants to query the MXs, it reuses the one already
recorded on the route if found (with proper refcounting). If it's
not there, it put itself in the waitq on this mx list.  If it's
the first one to enter the waitq, it triggers the call to lka.
Add the lka replies to the mxlist and activate the waitq when done,
resulting on all pending sessions to be resumed with the shared mx
list result.

The same mechanism is used to query the ptr name of an mx: only one
PTR lookup is done and all waiters are notified at once.